update shaders find same buffer slot for lines and points yay refactor shared data core update tests
