+++ src/DynamoCoreWpf/DynamoCoreWpf.csproj
@@ -303,7 +303,7 @@
     <Compile Include="ViewModels\Watch3D\DynamoGeometryModel3D.cs" />
     <Compile Include="ViewModels\Watch3D\DynamoLineGeometryModel3D.cs" />
     <Compile Include="ViewModels\Watch3D\DynamoPointGeometryModel3D.cs" />
-    <Compile Include="ViewModels\Watch3D\DynamoPointLineMaterialVariable.cs" />
+    <Compile Include="ViewModels\Watch3D\DynamoRenderCoreDataStore.cs" />
     <Compile Include="ViewModels\Watch3D\HelixWatch3DViewModel.cs" />
     <Compile Include="ViewModels\Watch3D\DefaultWatch3DViewModel.cs" />
     <Compile Include="ViewModels\Watch3D\IWatch3DViewModel.cs" />
