+++ src/DynamoCoreWpf/ViewModels/Watch3D/AttachedProperties.cs
@@ -328,12 +328,11 @@ namespace Dynamo.Wpf.ViewModels.Watch3D
         private static void HandlePointLinePropertyChange(GeometryModel3D pointLineGeom,
             DependencyPropertyChangedEventArgs args)
         {
+            //TODO should also work for lines as well, add type check
             if ((pointLineGeom as DynamoPointGeometryModel3D) != null)
             {
-                (pointLineGeom?.SceneNode?.RenderCore as DynamoPointLineCore).SetPropertyData(args);
-                (pointLineGeom as DynamoPointGeometryModel3D).BlendingFactor =
-                (pointLineGeom?.SceneNode?.RenderCore as DynamoPointLineCore).GenerateEnumFromState();
-            }
+                (pointLineGeom?.SceneNode?.RenderCore as DynamoPointLineCore).SetPropertyData(args,pointLineGeom as DynamoPointGeometryModel3D);
+            } 
         }
 
         /// <summary>
